RSDEV-1319-multiple-bump-version: refactoring for readable code
